+++ src/form/apo.jl
@@ -100,6 +100,16 @@ end
 
 
 
+""
+function constraint_switch_thermal_limit(pm::GenericPowerModel{T}, n::Int, c::Int, f_idx, rating) where T <: AbstractActivePowerFormulation
+    psw = var(pm, n, c, :psw, f_idx)
+
+    JuMP.lower_bound(psw) < -rating && JuMP.set_lower_bound(psw, -rating)
+    JuMP.upper_bound(psw) >  rating && JuMP.set_upper_bound(psw,  rating)
+end
+
+
+
 
 ""
 function constraint_storage_thermal_limit(pm::GenericPowerModel{T}, n::Int, c::Int, i, rating) where T <: AbstractActivePowerFormulation
@@ -132,5 +142,12 @@ function constraint_storage_on_off(pm::GenericPowerModel{T}, n::Int, i, pmin, pm
 
     JuMP.@constraint(pm.model, ps <= z_storage*pmax)
     JuMP.@constraint(pm.model, ps >= z_storage*pmin)
+end
+
+
+""
+function add_setpoint_switch_flow!(sol, pm::GenericPowerModel{T}) where T <: AbstractActivePowerFormulation
+    add_setpoint!(sol, pm, "switch", "psw", :psw, var_key = (idx,item) -> (idx, item["f_bus"], item["t_bus"]))
+    add_setpoint_fixed!(sol, pm, "switch", "qsw")
 end