+++ src/prob/test.jl
@@ -45,6 +45,211 @@ function _post_cl_opf(pm::GenericPowerModel)
 end
 
 
+"opf with fixed switches"
+function _run_sw_opf(file, model_constructor, optimizer; kwargs...)
+    return run_model(file, model_constructor, optimizer, _post_sw_opf; kwargs...)
+end
+
+""
+function _post_sw_opf(pm::GenericPowerModel)
+    variable_voltage(pm)
+    variable_generation(pm)
+    variable_switch_flow(pm)
+    variable_branch_flow(pm)
+    variable_dcline_flow(pm)
+
+    objective_min_fuel_and_flow_cost(pm)
+
+    constraint_model_voltage(pm)
+
+    for i in ids(pm, :ref_buses)
+        constraint_theta_ref(pm, i)
+    end
+
+    for i in ids(pm, :bus)
+        constraint_power_balance_shunt_switch(pm, i)
+    end
+
+    for i in ids(pm, :switch)
+        constraint_switch_state(pm, i)
+        constraint_switch_thermal_limit(pm, i)
+    end
+
+    for i in ids(pm, :branch)
+        constraint_ohms_yt_from(pm, i)
+        constraint_ohms_yt_to(pm, i)
+
+        constraint_voltage_angle_difference(pm, i)
+
+        constraint_thermal_limit_from(pm, i)
+        constraint_thermal_limit_to(pm, i)
+    end
+
+    for i in ids(pm, :dcline)
+        constraint_dcline(pm, i)
+    end
+end
+
+
+"opf with controlable switches"
+function _run_oswpf(file, model_constructor, optimizer; kwargs...)
+    return run_model(file, model_constructor, optimizer, _post_oswpf; ref_extensions=[ref_add_on_off_va_bounds!], solution_builder = _solution_osw!, kwargs...)
+end
+
+""
+function _post_oswpf(pm::GenericPowerModel)
+    variable_voltage(pm)
+    variable_generation(pm)
+
+    variable_switch_indicator(pm)
+    variable_switch_flow(pm)
+
+    variable_branch_flow(pm)
+    variable_dcline_flow(pm)
+
+    objective_min_fuel_and_flow_cost(pm)
+
+    constraint_model_voltage(pm)
+
+    for i in ids(pm, :ref_buses)
+        constraint_theta_ref(pm, i)
+    end
+
+    for i in ids(pm, :bus)
+        constraint_power_balance_shunt_switch(pm, i)
+    end
+
+    for i in ids(pm, :switch)
+        constraint_switch_on_off(pm, i)
+        constraint_switch_thermal_limit(pm, i)
+    end
+
+    for i in ids(pm, :branch)
+        constraint_ohms_yt_from(pm, i)
+        constraint_ohms_yt_to(pm, i)
+
+        constraint_voltage_angle_difference(pm, i)
+
+        constraint_thermal_limit_from(pm, i)
+        constraint_thermal_limit_to(pm, i)
+    end
+
+    for i in ids(pm, :dcline)
+        constraint_dcline(pm, i)
+    end
+end
+
+""
+function _solution_osw!(pm::GenericPowerModel, sol::Dict{String,<:Any})
+    add_setpoint_bus_voltage!(sol, pm)
+    add_setpoint_generator_power!(sol, pm)
+    add_setpoint_storage!(sol, pm)
+    add_setpoint_branch_flow!(sol, pm)
+    add_setpoint_dcline_flow!(sol, pm)
+    add_setpoint_switch_flow!(sol, pm)
+    add_setpoint_switch_status!(sol, pm)
+
+    add_dual_kcl!(sol, pm)
+    add_dual_sm!(sol, pm) # Adds the duals of the transmission lines' thermal limits.
+end
+
+
+
+"opf with controlable switches, node breaker"
+function _run_oswpf_nb(file, model_constructor, optimizer; kwargs...)
+    return run_model(file, model_constructor, optimizer, _post_oswpf_nb; ref_extensions=[ref_add_on_off_va_bounds!], solution_builder = _solution_osw_nb!, kwargs...)
+end
+
+""
+function _post_oswpf_nb(pm::GenericPowerModel)
+    variable_voltage_on_off(pm)
+    variable_generation(pm)
+
+    variable_switch_indicator(pm)
+    variable_switch_flow(pm)
+
+    variable_branch_indicator(pm)
+    variable_branch_flow(pm)
+    variable_dcline_flow(pm)
+
+    objective_min_fuel_and_flow_cost(pm)
+
+    constraint_model_voltage_on_off(pm)
+
+    for i in ids(pm, :ref_buses)
+        constraint_theta_ref(pm, i)
+    end
+
+    for i in ids(pm, :bus)
+        constraint_power_balance_shunt_switch(pm, i)
+    end
+
+    for i in ids(pm, :switch)
+        constraint_switch_on_off(pm, i)
+        constraint_switch_thermal_limit(pm, i)
+    end
+
+    for i in ids(pm, :branch)
+        constraint_ohms_yt_from_on_off(pm, i)
+        constraint_ohms_yt_to_on_off(pm, i)
+
+        constraint_voltage_angle_difference_on_off(pm, i)
+
+        constraint_thermal_limit_from_on_off(pm, i)
+        constraint_thermal_limit_to_on_off(pm, i)
+    end
+
+    for i in ids(pm, :dcline)
+        constraint_dcline(pm, i)
+    end
+
+    # link branch and switch variables
+    # branch indicator should only be zero iff,
+    # it has two swtiches and they are both zero
+    for (br, branch) in ref(pm, :branch)
+        bus_fr = branch["f_bus"]
+        bus_to = branch["t_bus"]
+        bus_switches = []
+
+        bus_arcs_sw_fr = ref(pm, :bus_arcs_sw, bus_fr)
+        if length(bus_arcs_sw_fr) == 1
+            for (sw,i,j) in bus_arcs_sw_fr
+                push!(bus_switches, sw)
+            end
+        end
+
+        bus_arcs_sw_to = ref(pm, :bus_arcs_sw, bus_to)
+        if length(bus_arcs_sw_to) == 1
+            for (sw,i,j) in bus_arcs_sw_to
+                push!(bus_switches, sw)
+            end
+        end
+
+        branch_z = var(pm, :branch_z, br)
+        switch_z = var(pm, pm.cnw, :z_switch)
+        for sw in bus_switches
+            JuMP.@NLconstraint(pm.model, branch_z >= switch_z[sw])
+        end
+        JuMP.@NLconstraint(pm.model, branch_z <= sum(switch_z[sw] for sw in bus_switches))
+    end
+end
+
+""
+function _solution_osw_nb!(pm::GenericPowerModel, sol::Dict{String,<:Any})
+    add_setpoint_bus_voltage!(sol, pm)
+    add_setpoint_generator_power!(sol, pm)
+    add_setpoint_storage!(sol, pm)
+    add_setpoint_branch_flow!(sol, pm)
+    add_setpoint_branch_status!(sol, pm)
+    add_setpoint_dcline_flow!(sol, pm)
+    add_setpoint_switch_flow!(sol, pm)
+    add_setpoint_switch_status!(sol, pm)
+
+    add_dual_kcl!(sol, pm)
+    add_dual_sm!(sol, pm) # Adds the duals of the transmission lines' thermal limits.
+end
+
+
 "opf with unit commitment, tests constraint_current_limit"
 function _run_uc_opf(file, model_constructor, solver; kwargs...)
     return run_model(file, model_constructor, solver, _post_uc_opf; solution_builder = _solution_uc!, kwargs...)